requirements.txt
fastapi
uvicorn
pydantic
prometheus-fastapi-instrumentator
databases
SQLAlchemy
psycopg2-binary
pytest
pytest-cov
mypy
flake8
slowapi
python-dotenv